telemetry-simulator/requirements.txt
asyncio-mqtt==0.16.2
asyncpg==0.29.0
neo4j==5.15.0
redis==5.0.1
numpy==1.26.2
pydantic==2.5.2
python-socketio[asyncio_client]==5.10.0
aiohttp==3.9.1
loguru==0.7.2
python-dotenv==1.0.0
PyYAML==6.0.1
httpx==0.25.2
websockets==12.0
pydantic-settings>=2.0.0